requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
asttokens==3.0.0
attrs==25.3.0
ccxt==4.4.90
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
circuitbreaker==2.1.3
comm==0.2.2
cryptography==44.0.3
dateparser==1.2.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
frozenlist==1.7.0
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
multidict==6.5.0
nest-asyncio==1.6.0
numpy==2.3.1
oci==2.154.2
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycares==4.9.0
pycparser==2.22
pycryptodome==3.23.0
Pygments==2.19.2
pyOpenSSL==24.3.0
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
